+++ server/services/favoriteLogoService.js
@@ -1,4 +1,5 @@
 const FavoriteLogo = require("../models/favoriteLogo.model");
+const { checkSQLWhereInputValid } = require("../utils/common");
 
 class FavoriteLogoService {
   static async getList() {
@@ -7,16 +8,28 @@ class FavoriteLogoService {
   }
 
   static async getListByUserId(user_id) {
+    if (!checkSQLWhereInputValid(user_id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     const list = await FavoriteLogo.where({ user_id }).fetchAll();
     return list;
   }
 
   static async getListByLogoId(logo_id) {
+    if (!checkSQLWhereInputValid(logo_id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     const list = await FavoriteLogo.where({ logo_id }).fetchAll();
     return list;
   }
 
   static async getByID(id) {
+    if (!checkSQLWhereInputValid(id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     const favorite = await FavoriteLogo.where({ id }).fetch();
     return favorite;
   }
@@ -29,6 +42,10 @@ class FavoriteLogoService {
   }
 
   static async updateById(id, payload) {
+    if (!checkSQLWhereInputValid(id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     let favorite = await FavoriteLogo.where({ id }).fetch();
     await favorite.save(payload);
     favorite = this.getByID(id);
@@ -36,6 +53,10 @@ class FavoriteLogoService {
   }
 
   static async deleteById(id) {
+    if (!checkSQLWhereInputValid(id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     await FavoriteLogo.where({ id }).destroy({
       require: false,
     });
