+++ server/services/leagueSeriesService.js
@@ -1,4 +1,5 @@
 const LeagueSeries = require("../models/leagueSeries.model");
+const { checkSQLWhereInputValid } = require("../utils/common");
 
 class LeagueSeriesService {
   static async getList() {
@@ -7,6 +8,10 @@ class LeagueSeriesService {
   }
 
   static async getListByUserId(userid) {
+    if (!checkSQLWhereInputValid(userid)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     const list = await LeagueSeries.query((qb) => {
       qb.join("leagues", "league_series.league_id", "=", "leagues.id");
       qb.where("leagues.userid", userid);
@@ -15,6 +20,10 @@ class LeagueSeriesService {
   }
 
   static async getByID(id) {
+    if (!checkSQLWhereInputValid(id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     const leagueSeries = await LeagueSeries.where({ id }).fetch();
     return leagueSeries;
   }
@@ -25,12 +34,20 @@ class LeagueSeriesService {
   }
 
   static async updateById(id, payload) {
+    if (!checkSQLWhereInputValid(id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     const leagueSeries = await LeagueSeries.where({ id }).fetch();
     await leagueSeries.save(payload);
     return leagueSeries;
   }
 
   static async deleteById(id) {
+    if (!checkSQLWhereInputValid(id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     await LeagueSeries.where({ id }).destroy({
       require: false,
     });
