+++ server/services/uploadService.js
@@ -1,4 +1,5 @@
 const Upload = require("../models/upload.model");
+const { checkSQLWhereInputValid } = require("../utils/common");
 
 class UploadService {
   static async getList() {
@@ -9,6 +10,10 @@ class UploadService {
   }
 
   static async getListByUserID(user_id) {
+    if (!checkSQLWhereInputValid(user_id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     const uploads = await Upload.where({
       user_id,
     }).fetchAll({
@@ -18,6 +23,10 @@ class UploadService {
   }
 
   static async getLegacyListByUserID(user_id) {
+    if (!checkSQLWhereInputValid(user_id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     const uploads = await Upload.where({
       user_id,
       legacy_mode: 1,
@@ -28,6 +37,10 @@ class UploadService {
   }
 
   static async getById(id) {
+    if (!checkSQLWhereInputValid(id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     const upload = await Upload.where({ id }).fetch({
       withRelated: ["user"],
     });
@@ -46,11 +59,25 @@ class UploadService {
   }
 
   static async deleteById(id) {
+    if (!checkSQLWhereInputValid(id)) {
+      throw new Error("SQL Injection attack detected.");
+    }
+
     await Upload.where({ id }).destroy({ require: false });
     return true;
   }
 
   static async deleteByMultiId(ids) {
+    if (!ids || !Array.isArray(ids) || !ids.length) {
+      throw new Error("Invalid input.");
+    }
+
+    for (let id of ids) {
+      if (!checkSQLWhereInputValid(id)) {
+        throw new Error("SQL Injection attack detected.");
+      }
+    }
+
     await Upload.where("id", "IN", ids).destroy({ require: false });
     return true;
   }
